Reorganized net protocol code
