+++ crates/net/src/lib.rs
@@ -1,7 +1,5 @@
 mod constants;
-pub mod encoding;
-pub mod packet_data;
-pub mod packet;
+pub mod protocol;
 pub mod utils;
 pub mod commands;
 pub mod connection;
