+++ crates/net/src/protocol/encoding.rs
